Rename mod skywalking_proto to proto.
